[SL-7] add footer component
