+++ src/App.js
@@ -6,106 +6,13 @@ import { useState, useEffect } from 'react'
 // components
 import ExerciseTables from './components/ExerciseTables'
 import NewExercise from './components/NewExercise'
-import QuickAddBtn from './components/buttons/QuickAddBtn'
+import Footer from './components/Footer'
 
 // animations
 import FadeIn from  './components/animations/FadeIn'
 
-// icons
-import { ReactComponent as DownloadSvg } from './images/icon/download.svg'
-import { ReactComponent as UploadSvg } from './images/icon/upload.svg'
-import pointerArrows from './images/btn-pointer-arrows.png'
-
-
 
 extendPrototype(localForage);
-
-var threeDaysAgo = new Date();
-threeDaysAgo.setDate(threeDaysAgo.getDate() - 3);
-
-var oneDayAgo = new Date();
-oneDayAgo.setDate(oneDayAgo.getDate() - 1);
-
-const Exercises2 = [
-  {
-    id: 4,
-    name: 'Chest Press',
-    sets: [
-      { weight: 100, reps: 8, side: null },
-      { weight: 100, reps: 6, side: null },
-      { weight: 85, reps: 7, side: null }
-    ],
-    massType: 'lb', // Lb or Kg
-    lengthType: 'm',
-    created: threeDaysAgo,
-    updated: null
-  },
-  {
-    id: 5,
-    name: 'Pull ups',
-    sets: [
-      { reps: 8 },
-      { reps: 7 },
-      { reps: 6 }
-    ],
-    massType: 'lb', // Lb or Kg
-    created: threeDaysAgo,
-    updated: null
-  },
-  {
-    id: 6,
-    name: 'Leg Press',
-    sets: [
-      { weight: 220, reps: 8 }, //
-      { weight: 220, reps: 8 },
-      { weight: 205, reps: 8 }
-    ],
-    massType: 'lb', // Lb or Kg
-    created: threeDaysAgo,
-    updated: null
-  }
-]
-
-const Exercises = [
-  {
-    id: 1,
-    name: 'Pec Fly Long Ass Name',
-    sets: [
-      { weight: 100, reps: 8, side: null }, //
-      { weight: 100, reps: 6, side: null },
-      { weight: 100, reps: 4, side: null },
-      { weight: 100, reps: 3, side: null },
-      { weight: 100, reps: 7, side: null }
-    ],
-    massType: 'lb', // Lb or Kg
-    lengthType: 'm',
-    created: oneDayAgo,
-    updated: null
-  },
-  {
-    id: 2,
-    name: 'Torso Rotate',
-    sets: [
-      { weight: 155, reps: 5, side: 'left' },
-      { weight: 155, reps: 7, side: 'right' }
-    ],
-    massType: 'lb', // Lb or Kg
-    created: oneDayAgo,
-    updated: null
-  },
-  {
-    id: 3,
-    name: 'Abs crunch',
-    sets: [
-      { reps: 8 },
-      { reps: 8 },
-      { reps: 7 }
-    ],
-    massType: 'lb', // Lb or Kg
-    created: oneDayAgo,
-    updated: null
-  }
-]
 
 function App() {
   localForage.config({
@@ -179,7 +86,6 @@ function App() {
   }
 
   /* === CRUD functions === */
-
   const loadExercises = (sessions, exercises) => {
     let populatedSession = sessions;
 
@@ -260,85 +166,8 @@ function App() {
     await localForage.setItem(exercise.id.toString(), exercise);
   }
 
-  const toggleNewExercise = () => {
-    return async() => {
-      await setNewExerciseMode(!newExerciseMode)
-    }
-  }
-
-  const downloadJson = async (fileName, contentType) => {
-    const dbItems = await localForage.getItems();
-    const exercisesArray = Object.values(dbItems);
-
-    let a = document.createElement("a");
-    const content = JSON.stringify(exercisesArray, null, 2);
-    console.log('content', content);
-    var file = new Blob([content], {type: 'text/plain'});
-    a.href = URL.createObjectURL(file);
-
-    a.download = `sess-log-exercises-${exercisesArray.length}.json`;
-    a.click();
-  }
-
-  const openUploader = () => {
-    let uploadInput = document.getElementById('uploadJsonInput');
-    uploadInput.click();
-  }
-
-  const uploadJson = async (e) => {
-    let promises =[];
-
-    try {
-      const fileReader = new FileReader();
-      fileReader.readAsText(e.target.files[0], "UTF-8");
-      fileReader.onload = e => {
-        let exercises = JSON.parse(e.target.result);
-
-        exercises.forEach(async (exercise) => {
-          console.log('exercise', exercise);
-          promises.push(localForage.setItem(exercise.id.toString(), exercise));
-        });
-      };
-    } catch (error) {
-      console.log('error 1', error);
-      alert('There was an error uploading your file. Please try again.')
-    }
-
-    try {
-        // clear db
-        await localForage.clear();
-        // populate w/ new data
-        await Promise.all(promises);
-        // load to state
-        let newSessions = await loadSessions();
-        await setSessions(newSessions);
-        console.log('')
-      } catch (error) {
-        console.log('error 2', error);
-        alert('There was an error processing your file. Please check its content.');
-      }
-  };
-
-  const circleBtnClass = `
-    rounded-full
-    bg-pink
-    flex
-    h-btn
-    items-center
-    justify-center
-    w-btn
-    mr-s
-  `;
-
-  const btnArrows = (
-    <div className="absolute flex-1 w-fit btn-arrow-pointers">
-      <img className="w-7/10" src={pointerArrows} alt="Arrows pointing to action buttons" />
-    </div>
-  );
-
   return (
     <div className="flex justify-center items-center sm:w-100 md:w-1/3 md:m-auto">
-      <input id="uploadJsonInput" type="file" className="hidden" onChange={uploadJson}/>
       <div className="relative h-screen flex-1 flex flex-col">
         <FadeIn isVisible={newExerciseMode} className="sticky top-0">
           <NewExercise
@@ -350,16 +179,15 @@ function App() {
         <section className="flex-1 flex flex-col">
           <ExerciseTables sessions={sessions} />
         </section>
-        <div className="sticky self-end mr-m bottom-m flex items-end">
-          { sessions.length > 0 ? null : btnArrows }
-          <button className={circleBtnClass} onClick={openUploader}>
-            <UploadSvg className="w-xm h-xm stroke-white"  />
-          </button>
-          <button className={circleBtnClass} onClick={downloadJson}>
-            <DownloadSvg className="w-xm h-xm stroke-white" />
-          </button>
-          <QuickAddBtn closeMode={newExerciseMode} onClick={toggleNewExercise()} />
-        </div>
+
+        <Footer
+          emptySessions={sessions.length < 1}
+          newExerciseMode={newExerciseMode}
+          setNewExerciseMode={setNewExerciseMode}
+          loadSessions={loadSessions}
+          setSessions={setSessions}
+        />
+
       </div>
     </div>
   );
